+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/RemoveWildcardImports.java
@@ -1,0 +1,13 @@
+package com.diffplug.spotless.maven.java;
+
+import com.diffplug.spotless.FormatterStep;
+import com.diffplug.spotless.java.RemoveWildcardImportsStep;
+import com.diffplug.spotless.maven.FormatterStepConfig;
+import com.diffplug.spotless.maven.FormatterStepFactory;
+
+public class RemoveWildcardImports implements FormatterStepFactory {
+	@Override
+	public FormatterStep newFormatterStep(FormatterStepConfig config) {
+		return RemoveWildcardImportsStep.create();
+	}
+}