+++ plugin-maven/CHANGES.md
@@ -3,6 +3,7 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+* Added `<removeWildcardImports />` step for Java, removing wildcard import statements. ([#649](https://github.com/diffplug/spotless/issues/649))
 
 ## [2.44.5] - 2025-05-27
 ### Changed